Add support for the internal TFTP server
This commit makes it possible to use `-t` without a path
to start the internal TFTP server to feed the image to the u-boot.

The root rights are required to open port 69 on the local host.

There are two options available:
```
./image_flasher.py -t <path_to_external_tftp_server> <path_to_image_file>
```
or
```
./image_flasher.py -t -- <path_to_image_file>
```
Last option uses `--` to separate the image file from the `-t` option.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>