Add the possibility to set IP in the u-boot
In case, when the network differs from default assumptions (192.168.1.*),
we need to specify the IP address of the board, the host, or both.

This commit adds the possibility to specify `--serverip` for host
and `--ipaddr` for the board.
If not provided then u-boot will use settings already available in
the environment.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>